Restructured this for Godot 3.1
